admin_fuzzer: fix too large param to mem stats size setter
Reviewed-on: https://gerrit.libreoffice.org/c/online/+/92568
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>

+++ wsd/Admin.cpp
@@ -222,7 +222,7 @@ void AdminSocketHandler::handleMessage(const std::vector<char> &payload)
             const std::string settingName = setting[0];
             if (settingName == "mem_stats_size")
             {
-                if (settingVal != std::stoi(model.query(settingName)))
+                if (settingVal != std::stol(model.query(settingName)))
                 {
                     model.setMemStatsSize(settingVal);
                 }
